+++ chaos_host/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_07_25_221908) do
+ActiveRecord::Schema[8.1].define(version: 2026_07_30_213054) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -114,6 +114,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_07_25_221908) do
     t.datetime "cancel_requested_at"
     t.datetime "created_at", null: false
     t.string "definitions_digest"
+    t.json "definitions_snapshot"
     t.string "failure_reason"
     t.datetime "finished_at"
     t.integer "index", null: false